requirements.txt
streamlit
librosa
matplotlib
numpy
torch
torchvision
pillow
